added dockerfiles

+++ src/main/java/ca/jrvs/apps/trading/model/domain/Position.java
@@ -15,7 +15,7 @@ public class Position {
   @JsonProperty("accountId")
   private Integer accountId;
   @JsonProperty("position")
-  private Integer position;
+  private Long position;
   @JsonProperty("ticker")
   private String ticker;
 
@@ -35,16 +35,16 @@ public class Position {
   }
 
   @JsonProperty("position")
-  public Integer getPosition() {
+  public Long getPosition() {
     return position;
   }
 
   @JsonProperty("position")
-  public void setPosition(Integer position) {
+  public void setPosition(Long position) {
     this.position = position;
   }
 
-  public Position withPosition(Integer position) {
+  public Position withPosition(Long position) {
     this.position = position;
     return this;
   }
